+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * tabChange event on tabGroup widget, tabWillChange events on the tab manager
 * URL argument template using format ?{key}
 * Deprecated global validateEmail() in favor of widget.validate.email()
+* Move (global) jquery plugins (sortlist, destroyed event) from widget-util to widget-core
 
 
 0.7.2 (2016.03.10)