+++ CHANGELOG.md
@@ -1,11 +1,15 @@
-0.7.3 (unreleased)
+0.8.0 (unreleased)
+--
+* EXPERIMENTAL - styles
+
+
+0.7.3 (2016.03.19)
 --
 * tabChange event on tabGroup widget, tabWillChange events on the tab manager
 * URL argument template using format ?{key}
 * Deprecated global validateEmail() in favor of widget.validate.email()
 * Move (global) jquery plugins (sortlist, destroyed event) from widget-util to widget-core
 * EXPERIMENTAL - form  and input field validation
-* EXPERIMENTAL - styles
 
 
 0.7.2 (2016.03.10)